+++ static/mail/2022/dockercon-2022-recap.html
@@ -711,7 +711,7 @@
                         
                         <td valign="top" class="mcnTextContent" style="padding-top: 0;padding-right: 18px;padding-bottom: 9px;padding-left: 18px;mso-line-height-rule: exactly;-ms-text-size-adjust: 100%;-webkit-text-size-adjust: 100%;word-break: break-word;color: #0e2117;font-family: 'Qanelas Soft', 'Helvetica Neue', Helvetica, Arial, sans-serif;font-size: 15px;line-height: 150%;text-align: left;">
                         
-                            HashiCorp’s Consul service mesh provides unique functionality to its users. Meshery's visual designer, <a href="https://layer5.io/service-mesh-management/meshmap" target="_blank" style="mso-line-height-rule: exactly;-ms-text-size-adjust: 100%;-webkit-text-size-adjust: 100%;color: #00b39f;font-weight: normal;text-decoration: none;">MeshMap</a>, allows developers to visually configure and deploy Consul-based deployments and workloads, allowing them to fully understand Consul’s differientiated capabilities.<br>
+                            HashiCorp’s Consul service mesh provides unique functionality to its users. Meshery's visual designer, <a href="https://layer5.io/service-mesh-management/kanvas" target="_blank" style="mso-line-height-rule: exactly;-ms-text-size-adjust: 100%;-webkit-text-size-adjust: 100%;color: #00b39f;font-weight: normal;text-decoration: none;">Kanvas</a>, allows developers to visually configure and deploy Consul-based deployments and workloads, allowing them to fully understand Consul’s differientiated capabilities.<br>
 <br>
 The <a href="https://layer5.io/docker-extension-meshery" target="_blank" style="mso-line-height-rule: exactly;-ms-text-size-adjust: 100%;-webkit-text-size-adjust: 100%;color: #00b39f;font-weight: normal;text-decoration: none;">Meshery extension’s</a> ability to import Docker Compose apps, transform them to Kubernetes applications, and deploy them on Consul service meshes is a powerful enabler for microservices developers, who need to dev, test, and deploy their modern applications in context of and compatibility with Consul service mesh.&nbsp;
                         </td>
@@ -942,7 +942,7 @@ The <a href="https://layer5.io/docker-extension-meshery" target="_blank" style="
     background-color: rgba(0,179,159,10%);
     ">
 <div style="text-align: center;">
-<p style="text-align: center;margin: 10px 0;padding: 0;mso-line-height-rule: exactly;-ms-text-size-adjust: 100%;-webkit-text-size-adjust: 100%;color: #0e2117;font-family: 'Qanelas Soft', 'Helvetica Neue', Helvetica, Arial, sans-serif;font-size: 15px;line-height: 150%;">The new <a href="https://layer5.io/docker-extension-meshery" target="_blank" style="mso-line-height-rule: exactly;-ms-text-size-adjust: 100%;-webkit-text-size-adjust: 100%;color: #00b39f;font-weight: normal;text-decoration: none;">Meshery Docker Extension</a> brings Layer5 <a href="https://layer5.io/service-mesh-management/meshmap" target="_blank" style="mso-line-height-rule: exactly;-ms-text-size-adjust: 100%;-webkit-text-size-adjust: 100%;color: #00b39f;font-weight: normal;text-decoration: none;">MeshMap</a>, the world's only visual designer for Kubernetes and service mesh deployments, to the desktop of millions of developers. Prepare for a demonstration&nbsp;of the capabilities of <a href="https://layer5.io/service-mesh-management/meshmap" target="_blank" style="mso-line-height-rule: exactly;-ms-text-size-adjust: 100%;-webkit-text-size-adjust: 100%;color: #00b39f;font-weight: normal;text-decoration: none;">MeshMap</a>.</p>
+<p style="text-align: center;margin: 10px 0;padding: 0;mso-line-height-rule: exactly;-ms-text-size-adjust: 100%;-webkit-text-size-adjust: 100%;color: #0e2117;font-family: 'Qanelas Soft', 'Helvetica Neue', Helvetica, Arial, sans-serif;font-size: 15px;line-height: 150%;">The new <a href="https://layer5.io/docker-extension-meshery" target="_blank" style="mso-line-height-rule: exactly;-ms-text-size-adjust: 100%;-webkit-text-size-adjust: 100%;color: #00b39f;font-weight: normal;text-decoration: none;">Meshery Docker Extension</a> brings Layer5 <a href="https://layer5.io/kanvas" target="_blank" style="mso-line-height-rule: exactly;-ms-text-size-adjust: 100%;-webkit-text-size-adjust: 100%;color: #00b39f;font-weight: normal;text-decoration: none;">Kanvas</a>, the world's only visual designer for Kubernetes and service mesh deployments, to the desktop of millions of developers. Prepare for a demonstration&nbsp;of the capabilities of <a href="https://layer5.io/service-mesh-management/" target="_blank" style="mso-line-height-rule: exactly;-ms-text-size-adjust: 100%;-webkit-text-size-adjust: 100%;color: #00b39f;font-weight: normal;text-decoration: none;">Kanvas</a>.</p>
 </div>
 </div>
